feat: Share UploadedFileCard component
